% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blue2dash.R
\name{blue2dash}
\alias{blue2dash}
\title{Takes the output from the blue function (recently, not old outputs!!) and reforms them into the new data required for the shiny dashboard}
\usage{
blue2dash(input_data_dir, output_dir, new_visitor_location = NULL)
}
\arguments{
\item{input_data_dir}{This is where the blue_function outputs are stored. You MUST have had "extra_csv_outputs=TRUE" selected when you ran the blue function, or this will fail. You can check this by seeing if icm_conv is a dataset in this folder location}

\item{output_dir}{This is where you would like the outputs to be saved.}

\item{new_visitor_location}{If you have run new visitors analysis, this is where that dataset is saved (the full .csv location).}
}
\value{
Returns nothing, but saves the data down to the specified location
}
\description{
Takes the output from the blue function (recently, not old outputs!!) and reforms them into the new data required for the shiny dashboard
Takes the output from the blue function (recently, not old outputs!!) and reforms them into the new data required for the shiny dashboard
}
\examples{

#blue2dash(input_data_dir="Z:/OMD/DISNEY_DP/05.May2016/OUTPUT/",
#           output_dir = "Z:/GROUP/TIM/shiny/blue2dash_testing/")
}
