% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blue_function.r
\name{blue_function}
\alias{blue_function}
\title{Runs the blue process}
\usage{
blue_function(username, DPID, file_save_location, cost_file_location,
  output_dir, click_cutoff = 5, imp_cutoff = 250, revenue_per_conv = 0,
  revenue_option = c("revenue_per_conv", "revenue_in_costs",
  "first_conversion_revenue", "all_conversion_revenue"),
  extra_csv_outputs = TRUE, searchgateway_sub = NULL, coefs_out = FALSE)
}
\arguments{
\item{username}{The username for redshift}

\item{file_save_location}{This is where you would like the intermediate files (between red and blue processes) to be saved.}

\item{output_dir}{This is where you would like the csv tabs for the dashboard to be output. If this is the same location as the tableau currently points then you will only need to refresh the dashboard to update. Do not have any of these files open while running.}

\item{click_cutoff}{The number of clicks that is counted as too few for a sub channel to be left on its own. Defaults to 5}

\item{imp_cutoff}{The number of imps that is counted as too few for a sub channel to be left on its own. Defaults to 250}

\item{revenue_per_conv}{The revenue for each conversion.}

\item{revenue_option}{can be any of 'revenue_per_conv','revenue_in_costs','first_conversion_revenue', or 'all_conversion_revenue'. The first looks to the 'revenue_per_conv argument and multiplies the weight column by this number. The second take sthe LC revenue from the cost file (if there is a column called revenue there). The third and fourth options take the revenue from the NA table generated in the red/gold function, and either takes the first purchase in the DPR period, or all purchases in the DPR period as the revenue for that user. This is the most advanced method.'}

\item{extra_csv_outputs}{Defaults to TRUE, asks if you want to write out as csv's some important intermediate files. For very large DP runs this may be undesirable as it can be slow for large files.}

\item{searchgateway_sub}{These are the sub channels which are determined to belong as a gateway event, usually brand search. These are treated differently to normal events. must be supplied as a character vector.}

\item{coefs_out}{This defaults to FALSE, but if TRUE will return the coefficients fitted by the process.}

\item{cost_save_location}{This is where the cost file is located, which must have columns channel, sub_channel,date,impressions,clicks,media_cost}
}
\value{
Summary statistics of the run process. The data is saved down in the file_save_location rather than outputted to the console.
}
\description{
Runs the blue process locally on R
}
\examples{
#blue_function(username="tim_scholtes",file_save_location="Z:/GROUP/DPR/TIM_test",
cost_file_location="Z:/GROUP/DPR/TIM_test",
output_dir="Z:/GROUP/DPR/TIM_test",
click_cutoff = 5,imp_cutoff=250,revenue_option="all_conversion_revenue") # not run
}
