% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blue_function_NV.r
\name{blue_function_NV}
\alias{blue_function_NV}
\title{Runs the blue process}
\usage{
blue_function_NV(username, DPID, file_save_location, output_dir,
  click_cutoff = 5, imp_cutoff = 250, extra_csv_outputs = TRUE)
}
\arguments{
\item{username}{The username for redshift}

\item{file_save_location}{This is where you would like the intermediate files (between red and blue processes) to be saved.}

\item{output_dir}{This is where you would like the csv tabs for the dashboard to be output. If this is the same location as the tableau currently points then you will only need to refresh the dashboard to update. Do not have any of these files open while running.}

\item{click_cutoff}{The number of clicks that is counted as too few for a sub channel to be left on its own. Defaults to 5}

\item{imp_cutoff}{The number of imps that is counted as too few for a sub channel to be left on its own. Defaults to 250}

\item{extra_csv_outputs}{Defaults to TRUE, asks if you want to write out as csv's some important intermediate files. For very large DP runs this may be undesirable as it can be slow for large files.}

\item{revenue_per_conv}{The revenue for each conversion.}
}
\value{
Summary statistics of the run process. The data is saved down in the file_save_location rather than outputted to the console.
}
\description{
Runs a simplified blue process specifically for the New Visitors analysis
}
\examples{
#blue_function_NV(username="tim_scholtes",file_save_location="Z:/GROUP/DPR/TIM_test",
output_dir="Z:/GROUP/DPR/TIM_test",
click_cutoff = 5,imp_cutoff=250,revenue_per_conv=10) # not run
}
