% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks_function.R
\name{checks_function1}
\alias{checks_function1}
\title{Does checks in the red function for stuff that will cause errors}
\usage{
checks_function1(startDate, endDate, startPre, endPre, reupload_hierarchy,
  hierarchy_location)
}
\arguments{
\item{reupload_hierarchy}{}
}
\value{
Summary statistics of the run process. The data is saved down in the file_save_location rather than outputted to the console.
}
\description{
Does checks in the red function for stuff that will cause errors
}
