% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks_function.R
\name{checks_function_keyword_PHD}
\alias{checks_function_keyword_PHD}
\title{Does checks in the red function for stuff that will cause errors}
\usage{
checks_function_keyword_PHD(startDate, endDate, startPre, endPre,
  reupload_hierarchy, ppc_match, hierarchy_location, ppc_hierarchy_location,
  ppc_match_col)
}
\arguments{
\item{reupload_hierarchy}{}
}
\value{
Summary statistics of the run process. The data is saved down in the file_save_location rather than outputted to the console.
}
\description{
Does checks in the red function for stuff that will cause errors
}
