% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost_function.R
\name{cost_function}
\alias{cost_function}
\title{Creates the cost file}
\usage{
cost_function(username, pw, dbname = "reportinguk",
  dbHost = "54.171.118.113", networkId, advertiserIds, campaignIds, hierarchy,
  startDate, endDate, allcosts = FALSE)
}
\arguments{
\item{username}{The username for redshift}

\item{pw}{The password for that username}

\item{dbname}{Database name, defaults to "reportinguk"}

\item{dbHost}{Database host, defaults to "54.171.118.113"}

\item{advertiserIds}{A character vector of advertiser ids to select}

\item{campaignIds}{A character vector of campaign ids to select}

\item{hierarchy}{Hierarchy dataframe}

\item{startDate}{the start date of the DP in "YYYY-MM-DD" format}

\item{endDate}{the end date of the DP in "YYYY-MM-DD" format}
}
\value{
Cost file containing media_cost, clicks and impressions by date, channel and subchannel
}
\description{
Automates the process to get the media_cost by day, channel and subchannel
}
\examples{

cost_function(username="patricia_martin",pw="",networkId="5851",
advertiserIds= c(
"2326173",    "2326266",    "4304358",    "4705211",
"4705212",    "4705613",    "4705617",    "4705805",
"4705811",    "4705815",    "4705826"),
campaignIds=c(
"7941825",  "8574112",  "8561809",
"8515806",    "8590693",    "8551379",    "8684699",
"8612335",    "8516389",    "8522817",    "8591270",
"8565152",    "8514189",    "8590689",    "8462957"),
hierarchy = hierarchy_dataframe,
startDate="2015-04-01",endDate="2015-04-30") #not run

}
