% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luna.R
\name{luna}
\alias{luna}
\title{Find the length of the unique values in an object}
\usage{
luna(x)
}
\arguments{
\item{x}{- the data}
}
\value{
The count of unique values
}
\description{
Find the length of the unique values in an object
}
\examples{
y <- c(sample(1:50,20,replace=T))
luna(y)
}
