% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_log_constrained.R
\name{mle_logreg_constrained}
\alias{mle_logreg_constrained}
\title{Fit the logistic regression}
\usage{
mle_logreg_constrained(fmla, data, maxit = 500, trace = TRUE)
}
\arguments{
\item{fmla}{- a formula string with the model to fit}

\item{data}{- the data which matches the fmla}

\item{maxit}{- maximum number of iterations}

\item{trace}{- BOOLEAN. True by default. TRUE gives a trace of the iterations into the console, FALSE gives no trace.}
}
\value{
A list containing the coefficients and the deviance
}
\description{
This function will fit
summary statistics and transformation of the data
}
\examples{
fmla <- as.formula("conv ~ sub_ch_A_1 + sub_ch_B_1 + sub_ch_A_2 + sub_ch_B_2")
data <- data.frame(conv=sample(c(0,1),size=1e3,replace=T),
                      sub_ch_A_1 = sample(0:4,size=1e3,replace=T),
                      sub_ch_B_1 = sample(0:4,size=1e3,replace=T),
                      sub_ch_A_2 = sample(0:4,size=1e3,replace=T),
                      sub_ch_B_2 = sample(0:4,size=1e3,replace=T))
                      
out <- mle_logreg_constrained(fmla,data)
}
