% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_zero.R
\name{na.zero}
\alias{na.zero}
\title{Replace NA's with zeroes}
\usage{
na.zero(x)
}
\arguments{
\item{x}{- the data with NAs}
}
\value{
The data with zeroes instead
}
\description{
Replace NA's with zeroes
}
\examples{
y <- data.frame(a=sample(c(1,2,3,NA),10,replace=T),
                b = sample(c(1,2,3,4,NA),10,replace=T))
na.zero(y)
}
