% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paste_bar.R
\name{paste_bar}
\alias{paste_bar}
\title{paste together with "|"}
\usage{
paste_bar(x)
}
\arguments{
\item{x}{- the data}
}
\value{
The pasted data as a string
}
\description{
paste together with "|"
}
\examples{
y <- c(sample(LETTERS,20,replace=T))
paste_bar(y)

}
