% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paste_bar_dedupe.R
\name{paste_bar_dedupe}
\alias{paste_bar_dedupe}
\title{paste deduped values with "|"}
\usage{
paste_bar_dedupe(x)
}
\arguments{
\item{x}{- the data}
}
\value{
The pasted data as a string
}
\description{
Dedupes consecutive values and pastes together with "|"
}
\examples{
y <- c(sample(LETTERS[1:10],20,replace=T))
paste_bar_dedupe(y)

}
