% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paste_comma.R
\name{paste_comma}
\alias{paste_comma}
\title{pastes together objects with commas}
\usage{
paste_comma(...)
}
\arguments{
\item{x}{- the data}
}
\value{
The pasted data
}
\description{
paste together objects with commas
}
\examples{
y <- c(sample(LETTERS,20,replace=T))
paste_comma(y)

}
