% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/red_function_keyword_page_id.R
\name{red_function_keyword_page_id}
\alias{red_function_keyword_page_id}
\title{Runs the initial red process}
\usage{
red_function_keyword_page_id(username, pw, DPID, dbname = "annalectuk",
  dbHost = "54.77.118.78", agency, networkId, advertiserIds, campaignIds,
  activity_ids, activity_type = NULL, eventid.ign = NULL, startDate,
  endDate, startPre, endPre, join_type = "Left", sample_factor = 20,
  sample_seed = NULL, parallel = T, file_save_location,
  hierarchy_location = NULL, ppc_hierarchy_location = NULL,
  ppc_match = TRUE, ppc_match_col = c("paid_search_campaign",
  "paid_search_ad_group"))
}
\arguments{
\item{username}{The username for redshift}

\item{pw}{The password for that username}

\item{dbname}{Database name, defaults to "annalectuk"}

\item{dbHost}{Database host, defaults to "54.77.118.78"}

\item{agency}{What agency are you? eg "m2m", "omd", "mg", "phd"}

\item{advertiserIds}{A character vector of advertiser ids to select}

\item{campaignIds}{A character vector of campaign ids to select}

\item{activity_ids}{Which activity ids are we counting as conversions?}

\item{eventid.ign}{By default NULL. You can specify an eventid type you wish to exclude from the convertors.}

\item{startDate}{the start date of the DP in "YYYY-MM-DD" format}

\item{endDate}{the end date of the DP in "YYYY-MM-DD" format}

\item{startPre}{the start date of the warmup period for the DP in "YYYY-MM-DD" format}

\item{endPre}{the end date of the warmup period for the DP in "YYYY-MM-DD" format}

\item{join_type}{The join type between the clicks and impressions to the hierarchy. "left" for a left join. Any placementids not in the hierarchy file will be filled in with "Other". "inner" for an inner join. Only channels in the hierarchy file will be included in the analysis.}

\item{sample_factor}{The multiple of converters you want from the non_converters - defaults to 20. This may be too large for some large clients (could cause memory prblems)}

\item{parallel}{Defaults to TRUE, whether you want to run some processes in parallel on your machine. Set this to false if you want to leave some cores available for other processes. The parallel operations are not running throughout so should cause minimal disruption.}

\item{file_save_location}{This is where you would like the intermediate files (between red and blue processes) to be saved.}

\item{hierarchy_location}{Where the hierarchy is saved. Defaults to safe gibberish.}

\item{ppc_hierarchy_location}{Where the ppc_hierarchy is saved. Defaults to safe gibberish.}

\item{ppc_match}{are we matching ppc separately?}
}
\value{
Summary statistics of the run process. The data is saved down in the file_save_location rather than outputted to the console.
}
\description{
Runs the initial red process
}
\examples{

#red_function_keyword(username="tim_scholtes",pw="",
agency="m2m",networkId="5851",
advertiserIds= c(
"2326173",    "2326266",    "4304358",    "4705211",
"4705212",    "4705613",    "4705617",    "4705805",
"4705811",    "4705815",    "4705826"),
campaignIds=c(
"7941825",  "8574112",  "8561809",
"8515806",    "8590693",    "8551379",    "8684699",
"8612335",    "8516389",    "8522817",    "8591270",
"8565152",    "8514189",    "8590689",    "8462957",
"8590872",    "8676743",    "8591271",    "8495905",
"8684932",    "8590678",    "8640691",    "8405167",
"8622631",    "8462956",    "8512842",    "8676843"),
activity_sub_type="hpuks0",
startDate="2015-04-01",endDate="2015-04-30",
startPre="2015-03-18",endPre="2015-03-31",
parallel=T, reupload_hierarchy = T,
file_save_location="Z:/GROUP/DPR/TIM_test/",
hierarchy_location="Z:/GROUP/DPR/TIM_test/hierarchy.csv") # not run

}
