% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshift_cleanup.R
\name{redshift_cleanup}
\alias{redshift_cleanup}
\title{Deletes tables on Redshift created from the red_function() and red_gold_function()}
\usage{
redshift_cleanup(username, pw, dbname = "annalectuk",
  dbHost = "54.77.118.78", DPID, agency)
}
\arguments{
\item{username}{The username for redshift}

\item{pw}{The password for that username}

\item{dbname}{Database name, defaults to "annalectuk"}

\item{dbHost}{Database host, defaults to "54.77.118.78"}

\item{DPID}{Unique identifier for the DP run you wish to delete tables for}

\item{agency}{What agency are you? eg "m2m", "omd", "mg", "phd"}

\item{networkID}{The network ID}
}
\description{
Deletes tables on Redshift created from the red_function() and red_gold_function()
}
\examples{
#redshift_cleanup(username="tim_scholtes", pw="", DPID="testrun", agency="m2m")
}
