% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_local_test.R
\name{shiny_local_test}
\alias{shiny_local_test}
\title{Tests the shiny dashboard on preprepared datasets from blue2dash}
\usage{
shiny_local_test(dashboard_file_location = NULL,
  app_location = "//ukncsas01/DATASCIENCE/SHARED/DP/DPR_shinydashboard/")
}
\arguments{
\item{dashboard_file_location}{The full path of the start_data.RData file from blue2dash()}

\item{app_location}{Where is the app stored? This defaults to the current location, but if there is an alternative app that has been created you can point it here}
}
\value{
Returns nothing, but will launch the app.
}
\description{
Tests the shiny dashboard on preprepared datasets from blue2dash
This function will launch the app code from a shared location (not from the server) to run on the specified data, so you can test all the outputs are A-OK. If they're not, email tim.scholtes@annalect.com with the details.
}
\examples{

#shiny_local_test(
#	dashboard_file_location="Z:/GROUP/TIM/shiny/blue2dash_testing/HP/start_data.RData"
#	)
}
